Fixup hexdump and cleanup main

+++ hexdump.h
@@ -5,6 +5,6 @@
 #include "lcd.h"
 
 char nibble_to_printable_hex(uint8_t nibble);
-void byte_to_printable_hex(uint16_t byte);
+void byte_to_printable_hex(uint8_t byte);
 
 #endif
